+++ tests/Models/User.php
@@ -1,12 +1,12 @@
 <?php namespace GeneaLabs\LaravelGovernor\Tests\Models;
 
-use GeneaLabs\LaravelGovernor\Traits\Governable;
+use GeneaLabs\LaravelGovernor\Traits\Governing;
 use Illuminate\Notifications\Notifiable;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 
 class User extends Authenticatable
 {
-    use Governable;
+    use Governing;
     use Notifiable;
 
     protected $fillable = [